delete function working 70%
delete function working by removing the element from the DOM but still stores the element in the local storage array. When you refresh the extension all the packages get loaded from the local storage again which is ok because we're gonna be putting them into a history cache for later and letting the user delete them permanently at a later time.