delete function 80% done now
delete function now not only removed the element from the DOM but also removes the element from the array (without pushing changes.. WIP for that one). Created a function that will act as a delta sync (only sync deltas) to push to the new array. Basically my thought process here is remove the entire array, then push the new one and then sync.