re-factored sync logic
Refactored original logic for performing an initial storage sync. Changed performSync function into async where it will start Promise to finish the function and then await the resolution at the end. After the resolution finishes it will call the newDelete function which resolves an issue with the delete button event listener from trying to access DOM elements that were not yet rendered. Optimized how the DOM obtained the ID of the button as this.value.